requirements.txt
customtkinter
Pillow
urllib3